Move vector stuff into a new file.
