+++ Wrappers/Python/cil/framework/framework.py
@@ -26,6 +26,8 @@ import math
 import logging
 
 from Wrappers.Python.cil.framework import DataContainer, AcquisitionData
+from Wrappers.Python.cil.framework.misc_vectors import ComponentDescription, PositionVector, DirectionVector, \
+    PositionDirectionVector
 
 # check for the extension
 
@@ -325,92 +327,6 @@ class ImageGeometry(object):
                 raise ValueError('Value {} unknown'.format(value))
 
         return out
-    
-class ComponentDescription(object):
-    r'''This class enables the creation of vectors and unit vectors used to describe the components of a tomography system
-     '''
-    def __init__ (self, dof):
-        self._dof = dof
-
-    @staticmethod  
-    def create_vector(val):
-        try:
-            vec = numpy.array(val, dtype=numpy.float64).reshape(len(val))
-        except:
-            raise ValueError("Can't convert to numpy array")
-   
-        return vec
-
-    @staticmethod   
-    def create_unit_vector(val):
-        vec = ComponentDescription.create_vector(val)
-        dot_product = vec.dot(vec)
-        if abs(dot_product)>1e-8:
-            vec = (vec/numpy.sqrt(dot_product))
-        else:
-            raise ValueError("Can't return a unit vector of a zero magnitude vector")
-        return vec
-
-    def length_check(self, val):
-        try:
-            val_length = len(val)
-        except:
-            raise ValueError("Vectors for {0}D geometries must have length = {0}. Got {1}".format(self._dof,val))
-        
-        if val_length != self._dof:
-            raise ValueError("Vectors for {0}D geometries must have length = {0}. Got {1}".format(self._dof,val))
-
-    @staticmethod   
-    def test_perpendicular(vector1, vector2):
-        dor_prod = vector1.dot(vector2)
-        if abs(dor_prod) <1e-10:
-            return True
-        return False
-
-    @staticmethod   
-    def test_parallel(vector1, vector2):
-        '''For unit vectors only. Returns true if directions are opposite'''
-        dor_prod = vector1.dot(vector2)
-        if 1- abs(dor_prod) <1e-10:
-            return True
-        return False
-
-class PositionVector(ComponentDescription):
-    r'''This class creates a component of a tomography system with a position attribute
-     '''
-    @property
-    def position(self):
-        try:
-            return self._position
-        except:
-            raise AttributeError
-
-    @position.setter
-    def position(self, val):  
-        self.length_check(val)
-        self._position = ComponentDescription.create_vector(val)
-
-
-class DirectionVector(ComponentDescription):
-    r'''This class creates a component of a tomography system with a direction attribute
-     '''
-    @property
-    def direction(self):      
-        try:
-            return self._direction
-        except:
-            raise AttributeError
-
-    @direction.setter
-    def direction(self, val):
-        self.length_check(val)    
-        self._direction = ComponentDescription.create_unit_vector(val)
-
- 
-class PositionDirectionVector(PositionVector, DirectionVector):
-    r'''This class creates a component of a tomography system with position and direction attributes
-     '''
-    pass
 
 class Detector1D(PositionVector):
     r'''This class creates a component of a tomography system with position and direction_x attributes used for 1D panels
@@ -880,7 +796,7 @@ class Parallel3D(SystemConfiguration):
         if numpy.allclose(self.rotation_axis.position, self.detector.position): #points are equal so on ray path
             rotation_axis_centred = True
         else:
-            vec_a = ComponentDescription.create_unit_vector(self.detector.position - self.rotation_axis.position )
+            vec_a = ComponentDescription.create_unit_vector(self.detector.position - self.rotation_axis.position)
             rotation_axis_centred = ComponentDescription.test_parallel(self.ray.direction, vec_a)
 
         if not rays_perpendicular_detector or\
@@ -1115,7 +1031,7 @@ class Cone2D(SystemConfiguration):
         if numpy.allclose(self.rotation_axis.position, self.detector.position): #points are equal
             rotation_axis_centred = True
         else:
-            vec_b = ComponentDescription.create_unit_vector(self.detector.position - self.rotation_axis.position )
+            vec_b = ComponentDescription.create_unit_vector(self.detector.position - self.rotation_axis.position)
             rotation_axis_centred = ComponentDescription.test_parallel(vec_src2det, vec_b)
 
         if not principal_ray_centred:
@@ -1313,7 +1229,7 @@ class Cone3D(SystemConfiguration):
         if numpy.allclose(self.rotation_axis.position, self.detector.position): #points are equal
             rotation_axis_centred = True
         else:
-            vec_b = ComponentDescription.create_unit_vector(self.detector.position - self.rotation_axis.position )
+            vec_b = ComponentDescription.create_unit_vector(self.detector.position - self.rotation_axis.position)
             rotation_axis_centred = ComponentDescription.test_parallel(vec_src2det, vec_b)
 
         if not principal_ray_centred or\
